+++ main.go
@@ -23,6 +23,7 @@ import (
 var (
 	disableDNSTCPCheck bool
 	hosts              string
+	dnsService         string
 	namespace          string
 	port               string
 	service            string
@@ -32,6 +33,7 @@ var (
 func init() {
 	flag.BoolVar(&disableDNSTCPCheck, "disable-dns-tcp-check", false, "Disable DNS TCP check")
 	flag.StringVar(&hosts, "hosts", "giantswarm.io.,kubernetes.default.svc.cluster.local.", "DNS hosts to resolve")
+	flag.StringVar(&dnsService, "dns-service", "coredns", "Name of DNS service")
 	flag.StringVar(&namespace, "namespace", "monitoring", "Namespace of net-exporter service")
 	flag.StringVar(&port, "port", "8000", "Port of net-exporter service")
 	flag.StringVar(&service, "service", "net-exporter", "Name of net-exporter service")
@@ -93,6 +95,7 @@ func main() {
 
 			DisableTCPCheck: disableDNSTCPCheck,
 			Hosts:           splitHosts,
+			Service:         dnsService,
 		}
 
 		dnsCollector, err = dns.New(c)